Address review feedback for Issue #262
