+++ test/gemini-pr-review.test.js
@@ -9,7 +9,9 @@ import {
   extractReviewerResponseFromGemini,
   findExistingGeminiReviewComment,
   formatGeminiReviewComment,
+  normalizeMentionLogin,
   parseGeminiReviewComment,
+  resolveOperatorMention,
   resolveGeminiReviewTrigger
 } from "../src/core/index.js";
 
@@ -276,6 +278,19 @@ test("formatGeminiReviewComment can render a short operator milestone mention", 
 
   assert.equal(body.split("\n")[1], "@marushu VTDD milestone: review result changed.");
   assert.equal(body.includes("Recommended action: `approve`"), true);
+});
+
+test("shared mention normalization filters non-operator GitHub actors", () => {
+  assert.equal(normalizeMentionLogin("marushu"), "marushu");
+  assert.equal(normalizeMentionLogin("github-actions[bot]"), "");
+  assert.equal(normalizeMentionLogin("vtdd-codex-bot"), "");
+  assert.equal(normalizeMentionLogin("ghost"), "");
+  assert.equal(normalizeMentionLogin("app/vtdd-codex"), "");
+});
+
+test("shared operator mention resolution falls back to the first mentionable actor", () => {
+  assert.equal(resolveOperatorMention(["github-actions[bot]", "issue-owner", "pr-owner"]), "issue-owner");
+  assert.equal(resolveOperatorMention(["ghost", "unknown", "github-actions[bot]"]), "");
 });
 
 test("findExistingGeminiReviewComment locates prior marker comment", () => {
